Extracted Metadata Support
The bearer token bits were mixing an older version
of the RSocket protocol with a newer one, causing some
confusion on the server side.

There was also some funny business with the route, so
I found a snippet online that appears to be more reliable.

Otherwise, just for clarity, I also removed the serializers
since they aren't necessary for this snippet